[UPDATE] change mobile route and adding multer
